requirements.txt
aiogram==2.25.1
flask==2.3.2
google.generativeai==0.3.2